advanced-api-project/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-authtoken
django-filter